chore: update on PR comments
