+++ crates/experimental-api/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod exe;
 pub mod gen;
+pub mod proof_validation;
 pub mod rpc;
