+++ crates/experimental-api/src/rpc.rs
@@ -113,6 +113,39 @@ struct Context {
     node: Arc<RwLock<NodeData>>,
 }
 
+impl Context {
+    pub async fn get_proper_block_id(
+        &self,
+        block_id: BlockId,
+    ) -> std::result::Result<BlockId, jsonrpc::Error> {
+        let l1_block_num = self.node.read().await.l1_block_number as i64;
+        match block_id {
+            gen::BlockId::BlockNumber { block_number } => {
+                let mut block_num = l1_block_num;
+                let user_request_block_num = *block_number.as_ref();
+                if user_request_block_num < block_num {
+                    block_num = user_request_block_num;
+                }
+                Ok(BlockId::BlockNumber {
+                    block_number: BlockNumber::try_new(block_num)?,
+                })
+            }
+            gen::BlockId::BlockHash { block_hash } => {
+                Ok(gen::BlockId::BlockHash { block_hash })
+            }
+            _ => Ok(BlockId::BlockNumber {
+                block_number: BlockNumber::try_new(l1_block_num)?,
+            }),
+        }
+    }
+
+    pub async fn get_l1_root(
+        &self,
+    ) -> std::result::Result<Felt, jsonrpc::Error> {
+        Felt::try_new(&format!("0x{:x}", self.node.read().await.l1_state_root))
+    }
+}
+
 async fn handle_request(
     State(ctx): State<Context>,
     Json(req): Json<Request>,
@@ -299,27 +332,8 @@ impl gen::Rpc for Context {
         key: StorageKey,
         block_id: BlockId,
     ) -> std::result::Result<Felt, jsonrpc::Error> {
-        let l1_block_num = self.node.read().await.l1_block_number as i64;
-        let block_id = match block_id {
-            gen::BlockId::BlockNumber {
-                block_number: BlockNumber(user_request_block_num),
-            } => {
-                let mut block_num = l1_block_num;
-                if user_request_block_num < block_num {
-                    block_num = user_request_block_num;
-                }
-                BlockId::BlockNumber {
-                    block_number: BlockNumber::try_new(block_num)?,
-                }
-            }
-            _ => BlockId::BlockNumber {
-                block_number: BlockNumber::try_new(l1_block_num)?,
-            },
-        };
-        let l1_root = Felt::try_new(&format!(
-            "0x{:x}",
-            self.node.read().await.l1_state_root
-        ))?;
+        let block_id = self.get_proper_block_id(block_id).await?;
+        let l1_root = self.get_l1_root().await?;
 
         let result = self
             .client
@@ -431,5 +445,116 @@ impl gen::Rpc for Context {
 
     async fn version(&self) -> std::result::Result<String, jsonrpc::Error> {
         self.client.version().await
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::sync::Arc;
+
+    use beerus_core::client::NodeData;
+    use tokio::sync::RwLock;
+
+    use crate::rpc::{BlockHash, BlockId, BlockNumber, BlockTag, Felt};
+
+    use super::{client::Client, Context};
+
+    #[tokio::test]
+    async fn returns_latest_block_number() {
+        let context = Context {
+            client: Arc::new(Client::new("")),
+            url: "".to_string(),
+            node: Arc::new(RwLock::new(NodeData {
+                l1_block_number: 3,
+                ..Default::default()
+            })),
+        };
+        assert_eq!(
+            context
+                .get_proper_block_id(BlockId::BlockNumber {
+                    block_number: BlockNumber::try_new(27).unwrap()
+                })
+                .await
+                .unwrap(),
+            BlockId::BlockNumber {
+                block_number: BlockNumber::try_new(3).unwrap()
+            }
+        );
+    }
+
+    #[tokio::test]
+    async fn returns_historical_block_number() {
+        let context = Context {
+            client: Arc::new(Client::new("")),
+            url: "".to_string(),
+            node: Arc::new(RwLock::new(NodeData {
+                l1_block_number: 42,
+                ..Default::default()
+            })),
+        };
+        assert_eq!(
+            context
+                .get_proper_block_id(BlockId::BlockNumber {
+                    block_number: BlockNumber::try_new(27).unwrap()
+                })
+                .await
+                .unwrap(),
+            BlockId::BlockNumber {
+                block_number: BlockNumber::try_new(27).unwrap()
+            }
+        );
+    }
+
+    #[tokio::test]
+    async fn on_latest_and_pending_returns_latest_block_number() {
+        let context = Context {
+            client: Arc::new(Client::new("")),
+            url: "".to_string(),
+            node: Arc::new(RwLock::new(NodeData {
+                l1_block_number: 42,
+                ..Default::default()
+            })),
+        };
+        assert_eq!(
+            context
+                .get_proper_block_id(BlockId::BlockTag(BlockTag::Latest))
+                .await
+                .unwrap(),
+            BlockId::BlockNumber {
+                block_number: BlockNumber::try_new(42).unwrap()
+            }
+        );
+        assert_eq!(
+            context
+                .get_proper_block_id(BlockId::BlockTag(BlockTag::Pending))
+                .await
+                .unwrap(),
+            BlockId::BlockNumber {
+                block_number: BlockNumber::try_new(42).unwrap()
+            }
+        );
+    }
+
+    #[tokio::test]
+    async fn on_hash_block_returns_hash_block() {
+        let context = Context {
+            client: Arc::new(Client::new("")),
+            url: "".to_string(),
+            node: Arc::new(RwLock::new(NodeData {
+                l1_block_number: 42,
+                ..Default::default()
+            })),
+        };
+        assert_eq!(
+            context
+                .get_proper_block_id(BlockId::BlockHash {
+                    block_hash: BlockHash(Felt::try_new("0xabc").unwrap())
+                })
+                .await
+                .unwrap(),
+            BlockId::BlockHash {
+                block_hash: BlockHash(Felt::try_new("0xabc").unwrap())
+            }
+        );
     }
 }
